Sending and revieving messages

+++ data/src/main/java/ru/kima/intelligentchat/data/kobold/horde/model/ContributionDto.kt
@@ -6,6 +6,6 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class ContributionDto(
-    @SerialName("megapixelsteps") val megapixelsteps: Int = 0,
+    @SerialName("megapixelsteps") val megapixelsteps: Double = 0.0,
     @SerialName("tokens") val tokens: Int = 0
 )
